Update models.py
This commit is meant to add a VERY basic listing table for our current django/postgreSQL setup. This does not yet work with the rest of the backend (at least I don't think). It is just the table that we can edit to add to / remove to as we need later